改善了前端页面的信息展示， 改善了用户订单的详细信息 完成了一个后台管理系统

+++ src/main/java/com/mobilephone/seckill/vo/GoodsVo.java
@@ -1,16 +1,17 @@
 package com.mobilephone.seckill.vo;
 
 
+import com.fasterxml.jackson.annotation.JsonFormat;
 import com.mobilephone.seckill.domain.Goods;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
 public class GoodsVo extends Goods {
-    private Double seckillPrice;
+    private BigDecimal seckillPrice;
     private Integer stockCount;
     private Date startDate;
     private Date endDate;
-
 }
